Fix extracting templates in all locales
The special characters are causing grep to detect that the file as
binary under some locales. This causes it output something like Binary
file /usr/share/ldapscripts/runtime matches instead of extracting the
embedded template required for adding users, groups etc. This results in
failure. Adding -a option to grep fixes the issue.